Change: hide stream implementation for Vec<Serialize>
As a user of response does not need to know about the internals of the
Stream implementation and since it is a requirment for ok_stream it
moves it to response crate and changes the interface of ok_stream to
hand over a vec instead of stream.

This allows changes of the interface of JsonArrayStreamer without
enforcing user of ok_stream to change.